chore(api): organize files
