+++ apps/api/src/lib/enum.ts
@@ -1,3 +1,15 @@
+export enum PrismaError {
+  UNIQUE_CONSTRAINT_VIOLATION = "P2002",
+  INVALID_FOREIGN_KEY = "P2003",
+  RECORD_TO_UPDATE_NOT_FOUND = "P2025",
+  RECORD_NOT_FOUND = "P2005",
+}
+
+export enum Role {
+  User = "user",
+  Admin = "admin",
+}
+
 export enum PaymentMethod {
   AYAPAY,
   KBZPAY,
