+++ apps/api/src/services/auth/interfaces/active-user.interface.ts
@@ -1,8 +1,10 @@
+import { UserType } from "src/lib/types";
+
 export interface ActiveUserData {
   sub: string;
   name: string;
   email: string;
-  type: "user" | "admin";
+  type: UserType;
   iat: number;
   exp: number;
 }
