--- apps/api/src/types/roles.enum.ts
@@ -1,4 +1,0 @@
-export enum Role {
-  User = "user",
-  Admin = "admin",
-}
