--- apps/api/src/utils/constants.ts
@@ -1,6 +1,0 @@
-export const PrismaError = {
-  UNIQUE_CONSTRAINT_VIOLATION: "P2002",
-  INVALID_FOREIGN_KEY: "P2003",
-  RECORD_TO_UPDATE_NOT_FOUND: "P2025",
-  RECORD_NOT_FOUND: "P2005",
-};
